WIP: Move oneoffs to new osmesa.apps subproject
These moves were done with IntelliJ "Move Class (F6)"
refactor tool. Using this tool does the following:
- Updates package reference
- Updates any import references (mostly automatic)
- Removes unused import references
- Reorders import references to alphabetical

I had to manually update most package/import refs
in docstring comments.
